simplify git subpackage
